+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This release is not backwards compatible. For easy migration best upgrade first 
 
 * Added support for Django REST framework 3.12
 * Added support for Django 3.1
+* Added support for Python 3.9
 
 ### Removed